Refector chemistry/elements functions and add unit tests
Goal:

The goal of this set of changes is add unit tests for existing
functionality and pull code into a 'neucbot' package, which can be
imported into other projects (i.e., a neucbot web interface, which may
be housed in a separate repo).

Additionally, some performance improvements will be made in order to
speed up execution time.

To that extent, the following changes have been made in this commit:

- Refactors Scripts/chemistry.py to create the inverse map of Z to
  element symbol.
- Pulls getNaturalIsotopes and getAbundance functionality into shared
  neucbot/elements.py module and uses the pregenerated
neucbot/elements.json file, to prevent the need for splitting/reading
the same file multiple times per execution.
- Adds unit tests for the files mentioned above
- Adds a requirements.txt file so that project dependencies can be
  installed into a Python virtual environment (allowing easier startup
for first time users).